+++ core/src/main/scala/scalacache/serialization/BaseCodecs.scala
@@ -1,4 +1,4 @@
-package scalacache.serdes
+package scalacache.serialization
 
 /**
  * Primitive type Codec instances
