+++ core/src/test/scala/issue42/Issue42Spec.scala
@@ -3,7 +3,7 @@ package issue42
 import org.scalatest.{ FlatSpec, Matchers }
 
 import scala.util.Random
-import scalacache.serdes.JavaSerializationCodecs._
+import scalacache.serialization.JavaSerializationCodecs._
 
 class Issue42Spec extends FlatSpec with Matchers {
