+++ core/src/test/scala/sample/Sample.scala
@@ -1,7 +1,7 @@
 package sample
 
 import scalacache._
-import scalacache.serdes.JavaSerializationCodecs._
+import scalacache.serialization.JavaSerializationCodecs._
 import memoization._
 
 import scala.concurrent.Future
